moved and renamed discretizers
